refactor: recoil 상태 관리 폴더명 변경(recoil->store)
